Refactors HttpServer.
Adds Server abstract class.
Turns HttpServer into a subclass of Server.
Changes return type of Http.createServer() method to HttpServer.
Removes ErrorEventListener marker interface.
