+++ src/main/java/io/bsonntag/neddy/http/HttpServer.java
@@ -1,12 +1,9 @@
 
 package io.bsonntag.neddy.http;
 
-import io.bsonntag.neddy.http.events.HttpRequestListener;
-import io.bsonntag.neddy.http.events.HttpServerErrorListener;
-import io.netty.bootstrap.ServerBootstrap;
+import io.bsonntag.neddy.events.BiEventListener;
+import io.bsonntag.neddy.net.Server;
 import io.netty.bootstrap.ServerChannelFactory;
-import io.netty.channel.Channel;
-import io.netty.channel.ChannelFuture;
 import io.netty.channel.ChannelInitializer;
 import io.netty.channel.EventLoopGroup;
 import io.netty.channel.socket.ServerSocketChannel;
@@ -18,11 +15,7 @@ import java.util.function.Supplier;
  * @author Benjamim Sonntag <benjamimsonntag@gmail.com>
  * @version 22/mai/2015
  */
-public final class HttpServer {
-    
-    private final EventLoopGroup bossGroup;
-    private final EventLoopGroup workerGroup;
-    private final ServerBootstrap serverBootstrap;
+public final class HttpServer extends Server<HttpRequest, HttpResponse> {
     
     private final HttpEventBus eventBus;
 
@@ -30,58 +23,20 @@ public final class HttpServer {
             ServerChannelFactory<ServerSocketChannel> channelFactory,
             ChannelInitializer channelInitializer,
             HttpEventBus eventBus) {
-        this.bossGroup = eventLoopFactory.get();
-        this.workerGroup = eventLoopFactory.get();
-        
+        super(eventLoopFactory, channelFactory, channelInitializer);
         this.eventBus = eventBus;
-        
-        serverBootstrap = new ServerBootstrap()
-                .group(bossGroup, workerGroup)
-                .channelFactory(channelFactory)
-                .childHandler(channelInitializer);
     }
     
-    public HttpServer onRequest(HttpRequestListener listener) {
+    @Override
+    public HttpServer onRequest(BiEventListener<HttpRequest, HttpResponse> listener) {
         eventBus.onRequest(listener);
         return this;
     }
     
-    public HttpServer onServerError(HttpServerErrorListener listener) {
+    @Override
+    public HttpServer onServerError(BiEventListener<Throwable, HttpResponse> listener) {
         eventBus.onServerError(listener);
         return this;
-    }
-    
-    public void listen(int port) {
-        ChannelFuture bindFuture = serverBootstrap.bind(port);
-        System.out.println("Server running on port " + port);
-        afterBind(bindFuture);
-    }
-    
-    public void listen(int port, String host) {
-        ChannelFuture bindFuture = serverBootstrap.bind(host, port);
-        System.out.println("Server running on port " + port);
-        afterBind(bindFuture);
-    }
-    
-    private void afterBind(ChannelFuture bindFuture) {
-        try {
-            waitForShutdown(bindFuture.channel());
-        }
-        catch(InterruptedException ex) {
-            ex.printStackTrace();
-        }
-        finally {
-            finishShutdown();
-        }
-    }
-    
-    private void waitForShutdown(Channel channel) throws InterruptedException {
-        channel.closeFuture().sync();
-    }
-
-    private void finishShutdown() {
-        workerGroup.shutdownGracefully();
-        bossGroup.shutdownGracefully();
     }
     
 }
